requirements.txt
python-dotenv
streamlit
streamlit_markdown
openai
jupyter
pandas
groq
langchain
langchain_community
ollama
